testbeds/netconf/docker/ngsi-ld-cleaner/pyproject.toml
[tool.poetry]
name = "ngsi-ld-cleaner"
version = "1.0.0"
description = "NGSI-LD cleaner"
authors = ["GIROS-DIT UPM <gi.giros@upm.es>"]

[tool.poetry.dependencies]
python = "^3.9.5"
urllib3 = ">=1.25.3,<2.1.0"
ngsi_ld_client = { git = "https://github.com/giros-dit/python-ngsi-ld-client.git", branch = "1.7.1" }
pyyaml = "^6.0"
python-dateutil = ">=2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = ">=4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
